agents/requirements.txt
azure-identity>=1.17.0
azure-ai-projects>=1.0.0b5
requests>=2.32.3
